src: kw_env: Adjust functions to use a encoded PWD as a cache_build_path
This commit implements get_encoded_pwd function in the global context
for solving duplicated env_names in different kernel trees. For every
env inside a tree, it will be created a folder in the global context
with its encoded pwd. The proper env_name folder will be created inside
it.

Signed-off-by: Gabriel Geraldino <gabrielgeraldinosouza@gmail.com>
Co-developed-by: Nicolas Borsari <nicborsari@usp.br>
Signed-off-by: Nicolas Borsari <nicborsari@usp.br>
Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>