src: send_patch: Add cross-subsystem devs to all patches
When using git-send-email together with get_maintainers in a series,
this can lead to a situation where parts of the series are only sent to
the main developers responsible for the specific part of the series, and
the other parts of the series are not shared. This situation can make
the review harder since patchsets are usually made in a context, and
including all of the people involved in the series makes more sense.
Technically, kw already knows all those involved in the series since it
extracts this information to be included in the cover letter. In this
sense, this commit expands this behavior to include all the same people
from the cover letter in each patch of the series.

Closes: #1207
Tested-by: Mario Limonciello (AMD) <superm1@kernel.org>
Signed-off-by: Rodrigo Siqueira <siqueira@igalia.com>