src: plugins: kernel_install: grub: Address deploy failure on Fedora
When using Fedora installed on the SD card in a Steam Deck device, it
has two grub.cfg files: one in/boot and the other in/boot/efi. Kw was
not addressing this situation, which led to a deploy failure. This
commit now considers this case and gives preference to the grub.cfg from
the efi folder.

Signed-off-by: Rodrigo Siqueira <siqueira@igalia.com>